Move TestSuite/TestCase base classes
Move TestSuite/TestCase base classes from mechanisms to a general folder.

Update VendorDeps
